// src/enums/Cards.ts
export enum Cards {
    DEFAULT_CARD = "default_of_card.png",
    ACE_CLUBS = "ace_of_clubs.png",
    ACE_SPADES = "ace_of_spades.png",
    ACE_DIAMONDS = "ace_of_diamonds.png",
    ACE_HEARTS = "ace_of_hearts.png",
  
    TWO_CLUBS = "2_of_clubs.png",
    TWO_SPADES = "2_of_spades.png",
    TWO_DIAMONDS = "2_of_diamonds.png",
    TWO_HEARTS = "2_of_hearts.png",
  
    THREE_CLUBS = "3_of_clubs.png",
    THREE_SPADES = "3_of_spades.png",
    THREE_DIAMONDS = "3_of_diamonds.png",
    THREE_HEARTS = "3_of_hearts.png",
  
    FOUR_CLUBS = "4_of_clubs.png",
    FOUR_SPADES = "4_of_spades.png",
    FOUR_DIAMONDS = "4_of_diamonds.png",
    FOUR_HEARTS = "4_of_hearts.png",
  
    FIVE_CLUBS = "5_of_clubs.png",
    FIVE_SPADES = "5_of_spades.png",
    FIVE_DIAMONDS = "5_of_diamonds.png",
    FIVE_HEARTS = "5_of_hearts.png",
  
    SIX_CLUBS = "6_of_clubs.png",
    SIX_SPADES = "6_of_spades.png",
    SIX_DIAMONDS = "6_of_diamonds.png",
    SIX_HEARTS = "6_of_hearts.png",
  
    SEVEN_CLUBS = "7_of_clubs.png",
    SEVEN_SPADES = "7_of_spades.png",
    SEVEN_DIAMONDS = "7_of_diamonds.png",
    SEVEN_HEARTS = "7_of_hearts.png",
  
    EIGHT_CLUBS = "8_of_clubs.png",
    EIGHT_SPADES = "8_of_spades.png",
    EIGHT_DIAMONDS = "8_of_diamonds.png",
    EIGHT_HEARTS = "8_of_hearts.png",
  
    NINE_CLUBS = "9_of_clubs.png",
    NINE_SPADES = "9_of_spades.png",
    NINE_DIAMONDS = "9_of_diamonds.png",
    NINE_HEARTS = "9_of_hearts.png",
  
    TEN_CLUBS = "10_of_clubs.png",
    TEN_SPADES = "10_of_spades.png",
    TEN_DIAMONDS = "10_of_diamonds.png",
    TEN_HEARTS = "10_of_hearts.png",
  
    JACK_CLUBS = "jack_of_clubs.png",
    JACK_SPADES = "jack_of_spades.png",
    JACK_DIAMONDS = "jack_of_diamonds.png",
    JACK_HEARTS = "jack_of_hearts.png",
  
    QUEEN_CLUBS = "queen_of_clubs.png",
    QUEEN_SPADES = "queen_of_spades.png",
    QUEEN_DIAMONDS = "queen_of_diamonds.png",
    QUEEN_HEARTS = "queen_of_hearts.png",
  
    KING_CLUBS = "king_of_clubs.png",
    KING_SPADES = "king_of_spades.png",
    KING_DIAMONDS = "king_of_diamonds.png",
    KING_HEARTS = "king_of_hearts.png",
};

export enum HandTypes {
    ROYAL_FLUSH = "royal_flush",
    STRAIGHT_FLUSH = "straight_flush",
    FOUR_OF_A_KIND = "four_of_a_kind",
    FULL_HOUSE = "full_house",
    FLUSH = "flush",
    STRAIGHT = "straight",
    THREE_OF_A_KIND = "three_of_a_kind",
    TWO_PAIR = "two_pair",
    PAIR = "pair",
    HIGH_CARD = "high_card"
}